fix pull request notes
