+++ src/Orc.CsvTextEditor/Operations/AddColumnOperation.cs
@@ -1,82 +1,5 @@
-﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="AddColumnOperation.cs" company="WildGums">
-//   Copyright (c) 2008 - 2019 WildGums. All rights reserved.
-// </copyright>
-// --------------------------------------------------------------------------------------------------------------------
-
-
-namespace Orc.CsvTextEditor.Operations
+﻿namespace Orc.CsvTextEditor.Operations
 {
-    using Catel.Logging;
-
-    internal class QuoteColumnOperation : OperationBase
-    {
-        private static readonly ILog Log = LogManager.GetCurrentClassLogger();
-
-        #region Constructors
-        public QuoteColumnOperation(ICsvTextEditorInstance csvTextEditorInstance)
-            : base(csvTextEditorInstance)
-        {
-        }
-        #endregion
-
-        public override void Execute()
-        {
-            var location = CsvTextEditorInstance.GetLocation();
-            var startPosition = location.Column.Offset + location.Line.Offset;
-            var endPosition = startPosition + location.Column.Width;
-
-            var text = CsvTextEditorInstance.GetText();
-
-            var quotesRemoved = false;
-            if (TryRemoveQuoteFromPosition(endPosition - 2, text, out var outputText))
-            {
-                text = outputText;
-                quotesRemoved = true;
-            }
-
-            if (TryRemoveQuoteFromPosition(startPosition, text, out outputText))
-            {
-                text = outputText;
-                quotesRemoved = true;
-            }
-
-            var offsetDelta = -1;
-            if (!quotesRemoved)
-            {
-                var quoteStr = Symbols.Quote.ToString();
-                text = text.Insert(startPosition, quoteStr)
-                    .Insert(endPosition, quoteStr);
-
-                offsetDelta = 1;
-            }
-
-            CsvTextEditorInstance.SetText(text);
-            CsvTextEditorInstance.GotoPosition(location.Offset + offsetDelta);
-
-            Log.Debug($"{nameof(QuoteColumnOperation)} executed; quotes were {(quotesRemoved ? "removed" : "added")}");
-        }
-
-        private static bool TryRemoveQuoteFromPosition(int symbolPosition, string inputText, out string outputText)
-        {
-            outputText = inputText;
-            if (symbolPosition >= inputText.Length || symbolPosition < 0)
-            {
-                return false;
-            }
-
-            var startSymbol = inputText[symbolPosition];
-            if (!Equals(startSymbol, Symbols.Quote))
-            {
-                return false;
-            }
-
-            outputText = inputText.Remove(symbolPosition, 1);
-
-            return true;
-        }
-    }
-
     internal class AddColumnOperation : OperationBase
     {
         #region Constructors
@@ -94,7 +17,7 @@ namespace Orc.CsvTextEditor.Operations
                 CsvTextEditorInstance.InsertAtCaret(Symbols.Comma);
                 return;
             }
-            
+
             var location = CsvTextEditorInstance.GetLocation();
 
             var column = location.Column;
@@ -130,7 +53,7 @@ namespace Orc.CsvTextEditor.Operations
             var quoteStr = Symbols.Quote.ToString();
 
             var text = CsvTextEditorInstance.GetText();
-            
+
             text = text.Insert(startPosition, quoteStr);
             text = text.Insert(location.Offset + 1, Symbols.Comma.ToString());
             text = text.Insert(startPosition + location.Column.Width + 1, quoteStr);
